algunos cambios

+++ EV3/src/main/java/Tingeso/EV3/Entities/CarreraEntity.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CarreraEntity {
     @Id
-    private Long codigo;
+    private int codigo;
     private String nombre;
 
 }
